Mengasingkan kod teras dalam subcrate 'onc' Fixes #3
