+++ wasm-rs/Cargo.toml
@@ -1,3 +1,4 @@
+workspace = { members = ["onc"] }
 [package]
 name = "wasm-rs"
 version = "0.1.0"
@@ -10,7 +11,7 @@ crate-type = ["cdylib"]
 [dependencies]
 clap = { version = "4.5.4", features = ["cargo"] }
 itertools = "0.12.1"
-nom = "7.1.3"
+onc = { version = "0.1.0", path = "onc" }
 serde = { version = "1.0.197", features = ["derive"] }
 serde-wasm-bindgen = "0.6.5"
 toml = "0.8.11"
