+++ wasm-rs/onc/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "onc"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+itertools = "0.12.1"
+nom = "7.1.3"
+serde = "1.0.197"
