+++ wasm-rs/onc/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod phonotactics;
+
+pub use nom::IResult;
